添加"我的收藏"中收藏的歌手、视频、专辑,支持查看歌曲的MV,优化代码结构

+++ imooc_voice/src/main/java/com/imooc/imooc_voice/model/newapi/song/SongDetailBean.java
@@ -102,7 +102,7 @@ public class SongDetailBean {
         private int copyright;
         private int s_id;
         private long mark;
-        private int mv;
+        private long mv;
         private int mst;
         private int cp;
         private int rtype;
@@ -312,7 +312,7 @@ public class SongDetailBean {
             this.mark = mark;
         }
 
-        public int getMv() {
+        public long getMv() {
             return mv;
         }
